eh: Make the compression thread freezable to fix suspend death
Since kthreads aren't freezable by default, it's possible for the
compression thread to run after the system is frozen and thus after EH has
been suspended. This is a fatal error, because EH hardware can neither
finish its work nor be interacted with after it is put to sleep.

Symptoms of the compression thread running after EH is suspended range from
complete system hangs due to the compression thread waiting for EH to
finish a compression that got lost during suspend, to kernel panics and
non-secure watchdog bites when the compression thread attempts to read EH
registers after EH is powered off.

As such, the compression thread must not run after EH is suspended. Fix
this by making the compression thread freezable, so that it must enter the
refrigerator before EH can suspend. Participating in system-wide freezing
also guarantees the compression thread won't be unfrozen until after EH is
resumed.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>